(BUGFIX) Correct some attributes type of Contact and Location entities
For example the phone attribute being an integer was giving me mixed results
during a call to an api GET, during the fetch of the contact data the
api was giving me the wrong result (the data of a contact that doesn't
count), I've found out that was the integer the problem (probably the
ORM has a problem for the sqlite3 db).

I've changed also the zip_code to string because it could begin with 0
and the integer type doesnt support this value at the beginning of the
number.